fix/test: rename and merge testcodes.
